// frontend/src/app/apiClient/model/escortDocument.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */






export interface EscortDocument { 

    
    documentID: string;

    
    escortID: string;

    
    type: string;

    
    name: string;

    
    storageHash?: string;

    
    createDate: Date;

    
    modifyDate?: Date;

}